backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy[asyncio]
pydantic
pydantic-settings
python-multipart
# Auth
PyJWT
bcrypt<4.0.0
passlib[bcrypt]
